pyproject.toml
[project]
name = "colcon-worktree-manager"
version = "0.1.0"
description = "Git worktree integration with colcon for parallel ROS 2 development"
readme = "README.md"
license = "Apache-2.0"
authors = [
    { name = "Kotaro Yoshimoto", email = "pythagora.yoshimoto@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "catkin-pkg>=1.0",
    "pyyaml>=6.0",
]

[project.scripts]
cwm = "cwm.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cwm"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["-p", "no:launch_testing", "-p", "no:launch_pytest", "-p", "no:colcon_core"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
]
